[bugfix] fix comparison between signed and unsigned warning

See merge request dune-fem/dune-fem!185

+++ dune/fem/space/mapper/parallel.hh
@@ -53,7 +53,7 @@ namespace Dune
           if( n == 0 )
             return;
 
-          assert( n == baseMapper_.numEntityDofs( entity ) );
+          assert( n == static_cast< std::size_t >( baseMapper_.numEntityDofs( entity ) ) );
           baseMapper_.mapEachEntityDof( entity, [ this, &buffer ] ( int, auto index ) {
               assert( slaveDofs_.isSlave( index ) );
               buffer.read( mapping_[ index ] );
